==> 01. Positions and UDL extracter.ipynb <==
import pandas as pd
import numpy as np

data_pddf = pd.read_excel('./input_files/positions.xlsx')

data_pddf.columns

positions_pddf = data_pddf[['Member', 'CC', 'Delta']]
positions_pddf.head()

udl_values_df = data_pddf[['CC', 'UDL value']]
udl_values_df = udl_values_df.drop_duplicates('CC').set_index('CC').sort_index()
udl_values_df.head()

udl_pddf = data_pddf[['UDL', 'Nu', 'Coef']]
udl_pddf = udl_pddf.set_index('UDL').sort_index()
udl_pddf = udl_pddf.dropna()
udl_pddf.head()

udl_pddf['UDL value'] = udl_values_df['UDL value']
udl_pddf.head()

members_id = list(set(positions_pddf['Member'].values))
members_id.sort(key=lambda s: int(s[2:]))

assets_id = list(set(udl_values_df.index.values))
assets_id.sort()

df_positions_matrix = pd.DataFrame(index=assets_id, columns=members_id)

for member in members_id:
    tmp_pos = positions_pddf.loc[positions_pddf['Member'] == member]
    tmp_pos = tmp_pos[['CC', 'Delta']].set_index('CC')['Delta']
    
    df_positions_matrix[member] = tmp_pos
    
df_positions_matrix[np.isnan(df_positions_matrix)] = 0.

df_positions_matrix = df_positions_matrix.transpose()

df_positions_matrix.head()

import os

dir_ = './transformed_input'

if not os.path.exists(dir_):
    os.makedirs(dir_)

udl_pddf.to_csv(os.path.join(dir_, 'UDL.csv'))
df_positions_matrix.to_csv(os.path.join(dir_, 'positions.csv'))
