==> 02. Random losses generator.ipynb <==
NU_COPULA = 6.
N_SIM = 100000

import pandas as pd
import numpy as np

import os

dir_ = './transformed_input'

udls = pd.read_csv(os.path.join(dir_, 'UDL.csv'), header=0, index_col=0)
positions = pd.read_csv(os.path.join(dir_, 'positions.csv'), header=0, index_col=0)

udl_labels = udls.index.tolist()

## Random variables

correl_matrix = pd.read_excel('./input_files/correlation_matrix.xlsx')
correl_matrix.set_index('Pearson', inplace=True)

correl_matrix = correl_matrix[udl_labels]
correl_matrix = correl_matrix.transpose()[udl_labels]

covar_matrix = correl_matrix

from scipy.stats import multivariate_normal, chi2, t

gauss_rv = multivariate_normal(cov=covar_matrix).rvs(N_SIM).transpose()
chi2_rv = chi2(NU_COPULA).rvs(N_SIM)

mult_factor = np.sqrt(NU_COPULA / chi2_rv)

student_rvs = np.multiply(mult_factor, gauss_rv)
inv_student_rvs = t.cdf(student_rvs, NU_COPULA)

print inv_student_rvs.shape

gauss_rv = None
chi2_rv = None
mult_factor = None
student_rvs = None

losses_rvs = np.ndarray(inv_student_rvs.shape)
losses_rvs[:] = np.nan

udls_trans = udls.transpose()

for i, col in enumerate(udls_trans.columns):
    params = udls_trans[col]
    
    nu = params['Nu']
    coeff = params['Coef']
    val = params['UDL value']
    
    crt_ppf = t.ppf(inv_student_rvs[i, :], nu)
    
    losses_rvs[i, :] = -val * coeff * crt_ppf
    
losses_rvs

losses_df = pd.DataFrame(losses_rvs, index=udl_labels, columns=[r'$\omega_{%i}$' % i for i in xrange(N_SIM)])
losses_df

import os

dir_ = './simulations'

if not os.path.exists(dir_):
    os.makedirs(dir_)

losses_df.to_csv(os.path.join(dir_, 'nu_eq_%i.csv' % NU_COPULA))
