==> 05. Initial Margins.ipynb <==
import pandas as pd
import numpy as np

import os

NU = 50

dir_sim_ = './simulations'
dir_tr_input_ = './transformed_input'

udl_distrib = pd.read_csv(os.path.join(dir_sim_,'nu_eq_%i.csv' % NU), header=0, index_col=0)
members_pos = pd.read_csv(os.path.join(dir_tr_input_, 'positions.csv'), header=0, index_col=0)

loss_and_profit = members_pos.dot(udl_distrib).T
print loss_and_profit.shape

im = pd.DataFrame()

for percentile in [99., 99.7]:
    quantiles_1 = loss_and_profit.quantile(percentile / 100.)
    quantiles_2 = -loss_and_profit.quantile(1. - percentile / 100.)
    
    quantiles = np.maximum(quantiles_1, quantiles_2)
    im[percentile] = quantiles

dir_ = './results'

if not os.path.exists(dir_):
    os.makedirs(dir_)

im.to_csv(os.path.join(dir_, 'IM_nu_eq_%i.csv' % NU))
